Moved delete account to settings. Fixed Phone Verification Success
